Cowsay.jl

A Julia package that lets you use cowsay in your Julia programs!

Usage

Cowsay.cowsay — Function

cowsay(message::AbstractString; kwargs...)
cowthink(message::AbstractString; kwargs...)

Print an ASCII picture of a cow saying/thinking message

Arguments

message::AbstractString: Tell the cow what to say

Keywords

cow=default: Specify a particular function handle to print the ASCII art. See below for more details on what constitutes a valid function handle.

eyes::AbstractString="oo": A two-character string to be drawn in for the eyes. Not all cowfiles support this, though.

tongue::AbstractString=" ": A two-character string to be drawn in for the tongue. Not all cowfiles support this.

Example

julia> cowsay("Have you mooed today?")
 _______________________
< Have you mooed today? >
 -----------------------
        \   ^__^
         \  (oo)\_______
            (__)\       )\/\
                ||----w |
                ||     ||

julia> cowthink("Have I mooed today?")
 _____________________
( Have I mooed today? )
 ---------------------
        o   ^__^
         o  (oo)\_______
            (__)\       )\/\
                ||----w |
                ||     ||