Cowsay.jl

A Julia package that lets you use cowsay in your Julia programs!

Usage

Cowsay.cowsay — Function

cowsay(message::AbstractString; kwargs...)

Print an ASCII picture of a cow saying message

Arguments

message::AbstractString: Tell the cow what to say

Keywords

cow=default: Specify a particular function handle to print the ASCII art. See below for more details on what constitutes a valid function handle.

eyes::AbstractString="oo": A two-character string to be drawn in for the eyes. Not all cowfiles support this, though.

tongue::AbstractString=" ": A two-character string to be drawn in for the tongue. Not all cowfiles support this.

Making a cow function

The original cowsay used Perl scripts (called 'cowfiles') to allow for creating more ASCII cow art. Cowsay.jl uses Julia functions, instead. In order to be usable by Cowsay.cowsay, a cow function must

Take the correct arguments

The function must take three (3) keyword arguments of the form

eyes::AbstractString="oo"

tongue::AbstractString=" "

thoughts::AbstractString="\"

When drawing the cow artwork, you may then use the variables eyes in place of the eyes, tongue in place of the tongue, and thoughts in place of the speech ballon trail. Use of these variables in constructing the cow is optional (but makes the use of your cow function far more fun), but all three arguments must be present in the signature, regardless.

Return a string

The cow artwork must be returned from the function as a string. This is distinctly different from how the original cowsay modified the $the_cow variable.

Helpful hints for making cow functions

Include one function per file, with the extension .cow.jl

Do not indent within a .cow.jl file to better see the artwork

Make use of string literals (""") and string interpolation ($) to build the cow art

Be sure to escape backslashes (\) and dollar signs ($) within your artwork

Split the eyes variable to get individual left- and right-eye when creating large cow functions

Have fun!