CowSay

Documentation for CowSay.

CowSay.bubble

CowSay.cowsay

CowSay.creature_combine_bubble

CowSay.message_wrap

CowSay.bubble — Method

bubble(msg_v::Vector)

Creates a message bubble.

CowSay.cowsay — Function

cowsay(s::String, creature="cow")

The infamous cowsay will be printed using this function

CowSay.creature_combine_bubble — Method

creature_combine_bubble(bubble::String, creature::String)

Combine message bubble and ascii art of creature.

CowSay.message_wrap — Method

message_wrap(msg::String)

Wraps the message into a compact format for each line so it can be used for function bubble.