Cows

Examples of all the cowfiles available.

Bovine

Cowsay.default — Function

function default()

The original cowfile

Example

julia> cowsay("Moo", cow=Cowsay.default)
 _____
< Moo >
 -----
        \   ^__^
         \  (oo)\_______
            (__)\       )\/\
                ||----w |
                ||     ||

Cowsay.cower — Function

function cower()

A cowering cow

Example

julia> cowsay("Is it safe to come out yet?", cow=Cowsay.cower)
 _____________________________
< Is it safe to come out yet? >
 -----------------------------
     \
      \
        ,__, |    |
        (oo)\|    |___
        (__)\|    |   )\_
             |    |_w |  \
             |    |  ||   *

             Cower....

Cowsay.udder — Function

function udder()

The cow from a file called cow-n-horn, artist unknown.

Example

julia> cowsay("Milking time!", cow=Cowsay.udder)
 _______________
< Milking time! >
 ---------------
  \
   \    (__)
        o o\
       ('') \---------
          \           \
           |          |\
           ||---(  )_|| *
           ||    UU  ||
           ==        ==

Cowsay.vader — Function

function vader()

Cowth Vader, from geordan@csua.berkeley.edu

Example

julia> cowsay("Luke, I am your father!", cow=Cowsay.vader)
 _________________________
< Luke, I am your father! >
 -------------------------
        \    ,-^-.
         \   !oYo!
          \ /./=\.\______
               ##        )\/\
                ||-----w||
                ||      ||

               Cowth Vader

Cowsay.www — Function

function www()

A cow wadvertising the World Wide Web, from lim@csua.berkeley.edu

Example

julia> cowsay("My favorite site is MooTube", cow=Cowsay.www)
 _____________________________
< My favorite site is MooTube >
 -----------------------------
        \   ^__^
         \  (oo)\_______
            (__)\       )\/\
                ||--WWW |
                ||     ||

Mascots

Cowsay.blowfish — Function

function blowfish()

A blowfish.

julia> cowsay("Bloop, bloop", cow=Cowsay.blowfish)
 ______________
< Bloop, bloop >
 --------------
   \
    \
               |    .
           .   |L  /|
       _ . |\ _| \--+._/| .
      / ||\| Y J  )   / |/| ./
     J  |)'( |        ` F`.'/
   -<|  F         __     .-<
     | /       .-'. `.  /-. L___
     J \      <    \  | | O\|.-'
   _J \  .-    \/ O | | \  |F
  '-F  -<_.     \   .-'  `-' L__
 __J  _   _.     >-'  )._.   |-'
 `-|.'   /_.           \_|   F
   /.-   .                _.<
  /'    /.'             .'  `\
   /L  /'   |/      _.-'-\
  /'J       ___.---'\|
    |\  .--' V  | `. `
    |/`. `-.     `._)
       / .-.\
 VK    \ (  `\
        `.\

Cowsay.elephant — Function

function elephant()

An elephant out and about

Example

julia> cowsay("Toot!", cow=Cowsay.elephant)
 _______
< Toot! >
 -------
 \     /\  ___  /\
  \   // \/   \/ \\
     ((    O O    ))
      \\ /     \ //
       \/  | |  \/
        |  | |  |
        |  | |  |
        |   o   |
        | |   | |
        |m|   |m|

Other

Cowsay.bunny — Function

function bunny()

A cute little wabbit

Example

julia> cowsay("Look at that tortoise way off in the distance.", cow=Cowsay.bunny)
 ________________________________________________
< Look at that tortoise way off in the distance. >
 ------------------------------------------------
  \
   \   \
        \ /\
        ( )
      .( o ).

Cowsay.dragon_and_cow — Function

function dragon_and_cow()

A dragon smiting a cow, possible credit to kube@csua.berkeley.edu

Example

julia> cowsay("Roast beef: it's what's for dinner", cow=Cowsay.dragon_and_cow)
 ____________________________________
< Roast beef: it's what's for dinner >
 ------------------------------------
                       \                    ^    /^
                        \                  / \  // \
                         \   |\___/|      /   \//  .\
                          \  /O  O  \__  /    //  | \ \           *----*
                            /     /  \/_/    //   |  \  \          \   |
                            @___@`    \/_   //    |   \   \         \/\ \
                           0/0/|       \/_ //     |    \    \         \  \
                       0/0/0/0/|        \///      |     \     \       |  |
                    0/0/0/0/0/_|_ /   (  //       |      \     _\     |  /
                 0/0/0/0/0/0/`/,_ _ _/  ) ; -.    |    _ _\.-~       /   /
                             ,-}        _      *-.|.-~-.           .~    ~
            \     \__/        `/\      /                 ~-. _ .-~      /
             \____(oo)           *.   }            {                   /
             (    (--)          .----~-.\        \-`                 .~
             //__\\  \__ Ack!   ///.----..<        \             _ -~
            //    \\               ///-._ _ _ _ _ _ _{^ - - - - ~

Cowsay.dragon — Function

function dragon()

The Whitespace Dragon

Example

julia> cowsay("Roar!", cow=Cowsay.dragon)
 _______
< Roar! >
 -------
      \                    / \  //\
       \    |\___/|      /   \//  \\
            /0  0  \__  /    //  | \ \
           /     /  \/_/    //   |  \  \
           @_^_@'/   \/_   //    |   \   \
           //_^_/     \/_ //     |    \    \
        ( //) |        \///      |     \     \
      ( / /) _|_ /   )  //       |      \     _\
    ( // /) '/,_ _ _/  ( ; -.    |    _ _\.-~        .-~~~^-.
  (( / / )) ,-{        _      `-.|.-~-.           .~         `.
 (( // / ))  '/\      /                 ~-. _ .-~      .-~^-.  \
 (( /// ))      `.   {            }                   /      \  \
  (( / ))     .----~-.\        \-'                 .~         \  `. \^-.
             ///.----..>        \             _ -~             `.  ^-`  ^-_
               ///-._ _ _ _ _ _ _}^ - - - - ~                     ~-- ,.-~
                                                                  /.-~

Cowsay.elephant_in_snake — Function

function elephant_in_snake()

Do we need to explain this?

Example

julia> cowsay("Help!", cow=Cowsay.elephant_in_snake)
 _______
< Help! >
 -------
   \
    \              ....
           ........    .
          .            .
         .             .
.........              .......
..............................

Elephant inside ASCII snake

Cowsay.vader_koala — Function

function vader_koala

Another canonical koala?

Example

julia> cowsay("Luke, you are my joey!", cow=Cowsay.vader_koala)
 ________________________
< Luke, you are my joey! >
 ------------------------
   \
    \        .
     .---.  //
    Y|o o|Y//
   /_(i=i)K/
   ~()~*~()~
    (_)-(_)

     Darth
     Vader
     koala