Cows

Examples of all the cowfiles available.

Bovine

Cowsay.default — Function

function default()

The original cowfile

Example

julia> cowsay("Moo", cow=Cowsay.default)
 _____
< Moo >
 -----
        \   ^__^
         \  (oo)\_______
            (__)\       )\/\
                ||----w |
                ||     ||

Cowsay.cower — Function

function cower()

A cowering cow

Example

julia> cowsay("Is it safe to come out yet?", cow=Cowsay.cower)
 _____________________________
< Is it safe to come out yet? >
 -----------------------------
     \
      \
        ,__, |    |
        (oo)\|    |___
        (__)\|    |   )\_
             |    |_w |  \
             |    |  ||   *

             Cower....

Cowsay.moofasa — Function

function moofasa()

MOOfasa

Example

julia> cowsay("It's the circle of life", cow=Cowsay.moofasa)
 _________________________
< It's the circle of life >
 -------------------------
       \    ____
        \  /    \
          | ^__^ |
          | (oo) |______
          | (__) |      )\/\
           \____/|----w |
                ||     ||

             Moofasa

Cowsay.mutilated — Function

function mutilated()

A mutilated cow, from aspolito@csua.berkeley.edu

Example

julia> cowsay("I wouldn't have been so eager to make head cheese...", cow=Cowsay.mutilated)
 _______________________________________
/ I wouldn't have been so eager to make \
\ head cheese...                        /
 ---------------------------------------
       \   \_______
 v__v   \  \   O   )
 (oo)      ||----w |
 (__)      ||     ||  \/\

Cowsay.skeleton — Function

function skeleton()

This 'Scowleton' brought to you by one of {appel,kube,rowe}@csua.berkeley.edu

Example

julia> cowsay("All T-Bone, no steak", cow=Cowsay.skeleton)
 ______________________
< All T-Bone, no steak >
 ----------------------
          \      (__)
           \     /oo|
            \   (_"_)*+++++++++*
                   //I#\\\\\\\\I\
                   I[I|I|||||I I `
                   I`I'///'' I I
                   I I       I I
                   ~ ~       ~ ~
                     Scowleton

Cowsay.small — Function

function small()

A small cow, artist unknown

Example

julia> cowsay("A calf", cow=Cowsay.small)
 ________
< A calf >
 --------
       \   ,__,
        \  (..)____
           (__)    )\
              ||--|| *

Cowsay.supermilker — Function

function supermilker()

A cow being milked, probably from Lars Smith (lars@csua.berkeley.edu)

Example

julia> cowsay("Paying the bills", cow=Cowsay.supermilker)
 __________________
< Paying the bills >
 ------------------
  \   ^__^
   \  (oo)\_______        ________
      (__)\       )\/\    |Super |
          ||----W |       |Milker|
          ||    UDDDDDDDDD|______|

Cowsay.three_eyes — Function

function three_eyes()

A cow with three eyes, brought to you by dpetrou@csua.berkeley.edu

Example

julia> cowsay("The better to see you with...", cow=Cowsay.three_eyes)
 _______________________________
< The better to see you with... >
 -------------------------------
        \  ^___^
         \ (ooo)\_______
           (___)\       )\/\
                ||----w |
                ||     ||

Cowsay.udder — Function

function udder()

The cow from a file called cow-n-horn, artist unknown.

Example

julia> cowsay("Milking time!", cow=Cowsay.udder)
 _______________
< Milking time! >
 ---------------
  \
   \    (__)
        o o\
       ('') \---------
          \           \
           |          |\
           ||---(  )_|| *
           ||    UU  ||
           ==        ==

Cowsay.vader — Function

function vader()

Cowth Vader, from geordan@csua.berkeley.edu

Example

julia> cowsay("Luke, I am your father!", cow=Cowsay.vader)
 _________________________
< Luke, I am your father! >
 -------------------------
        \    ,-^-.
         \   !oYo!
          \ /./=\.\______
               ##        )\/\
                ||-----w||
                ||      ||

               Cowth Vader

Cowsay.www — Function

function www()

A cow wadvertising the World Wide Web, from lim@csua.berkeley.edu

Example

julia> cowsay("My favorite site is MooTube", cow=Cowsay.www)
 _____________________________
< My favorite site is MooTube >
 -----------------------------
        \   ^__^
         \  (oo)\_______
            (__)\       )\/\
                ||--WWW |
                ||     ||

Mascots

Cowsay.blowfish — Function

function blowfish()

A blowfish.

julia> cowsay("Bloop, bloop", cow=Cowsay.blowfish)
 ______________
< Bloop, bloop >
 --------------
   \
    \
               |    .
           .   |L  /|
       _ . |\ _| \--+._/| .
      / ||\| Y J  )   / |/| ./
     J  |)'( |        ` F`.'/
   -<|  F         __     .-<
     | /       .-'. `.  /-. L___
     J \      <    \  | | O\|.-'
   _J \  .-    \/ O | | \  |F
  '-F  -<_.     \   .-'  `-' L__
 __J  _   _.     >-'  )._.   |-'
 `-|.'   /_.           \_|   F
   /.-   .                _.<
  /'    /.'             .'  `\
   /L  /'   |/      _.-'-\
  /'J       ___.---'\|
    |\  .--' V  | `. `
    |/`. `-.     `._)
       / .-.\
 VK    \ (  `\
        `.\

Cowsay.elephant — Function

function elephant()

An elephant out and about

Example

julia> cowsay("Toot!", cow=Cowsay.elephant)
 _______
< Toot! >
 -------
 \     /\  ___  /\
  \   // \/   \/ \\
     ((    O O    ))
      \\ /     \ //
       \/  | |  \/
        |  | |  |
        |  | |  |
        |   o   |
        | |   | |
        |m|   |m|

Cowsay.tux — Function

function tux()

TuX (c) pborys@p-soft.silesia.linux.org.pl

Example

julia> cowsay("Talk is cheap. Show me the code.", cow=Cowsay.tux)
 __________________________________
< Talk is cheap. Show me the code. >
 ----------------------------------
   \
    \
        .--.
       |o_o |
       |:_/ |
      //   \ \
     (|     | )
    /'\_   _/`\
    \___)=(___/

Cows and friends

Cowsay.dragon_and_cow — Function

function dragon_and_cow()

A dragon smiting a cow, possible credit to kube@csua.berkeley.edu

Example

julia> cowsay("Roast beef: it's what's for dinner", cow=Cowsay.dragon_and_cow)
 ____________________________________
< Roast beef: it's what's for dinner >
 ------------------------------------
                       \                    ^    /^
                        \                  / \  // \
                         \   |\___/|      /   \//  .\
                          \  /O  O  \__  /    //  | \ \           *----*
                            /     /  \/_/    //   |  \  \          \   |
                            @___@`    \/_   //    |   \   \         \/\ \
                           0/0/|       \/_ //     |    \    \         \  \
                       0/0/0/0/|        \///      |     \     \       |  |
                    0/0/0/0/0/_|_ /   (  //       |      \     _\     |  /
                 0/0/0/0/0/0/`/,_ _ _/  ) ; -.    |    _ _\.-~       /   /
                             ,-}        _      *-.|.-~-.           .~    ~
            \     \__/        `/\      /                 ~-. _ .-~      /
             \____(oo)           *.   }            {                   /
             (    (--)          .----~-.\        \-`                 .~
             //__\\  \__ Ack!   ///.----..<        \             _ -~
            //    \\               ///-._ _ _ _ _ _ _{^ - - - - ~

Cowsay.mech_and_cow — Function

function mech_and_cow()

Example

julia> cowsay("Do you ever get that feeling...?", cow=Cowsay.mech_and_cow)
 __________________________________
< Do you ever get that feeling...? >
 ----------------------------------
   \                               ,-----.
   /                               |     |
   \                            ,--|     |-.
   /                     __,----|  |     | |
   \                   ,;::     |  `_____' |
   /                   `._______|    i^i   |
   \                            `----| |---'| .
   /                       ,-------._| |== ||//
   \                       |       |_|P`.  /'/
   /                       `-------' 'Y Y/'/'
   \                                 .==\ /_\
   ^__^                             /   /'|  `i
   (oo)\_______                   /'   /  |   |
   (__)\       )\/\            /'    /   |   `i
      ||----w||           ___,;`----'.___L_,-'`\__
      ||     ||          i_____;----\.____i""\____\

Cowsay.surgery — Function

function surgery()

A cow operation, artist unknown

Example

julia> cowsay("Removing the last bit of net wrap now", cow=Cowsay.surgery)
 _______________________________________
< Removing the last bit of net wrap now >
 ---------------------------------------
          \           \  /
           \           \/
               (__)    /\
               (oo)   O  O
               _\/_   //
         *    (    ) //
          \  (\\    //
           \(  \\    )
            (   \\   )   /\
  ___[\______/^^^^^^^\__/) o-)__
 |\__[=======______//________)__\
 \|_______________//____________|
     |||      || //||     |||
     |||      || @.||     |||
      ||      \/  .\/      ||
                 . .
                '.'.`

            COW-OPERATION

Other

Cowsay.bunny — Function

function bunny()

A cute little wabbit

Example

julia> cowsay("Look at that tortoise way off in the distance.", cow=Cowsay.bunny)
 ______________________________________
/ Look at that tortoise way off in the \
\ distance.                            /
 --------------------------------------
  \
   \   \
        \ /\
        ( )
      .( o ).

Cowsay.dragon — Function

function dragon()

The Whitespace Dragon

Example

julia> cowsay("Roar!", cow=Cowsay.dragon)
 _______
< Roar! >
 -------
      \                    / \  //\
       \    |\___/|      /   \//  \\
            /0  0  \__  /    //  | \ \
           /     /  \/_/    //   |  \  \
           @_^_@'/   \/_   //    |   \   \
           //_^_/     \/_ //     |    \    \
        ( //) |        \///      |     \     \
      ( / /) _|_ /   )  //       |      \     _\
    ( // /) '/,_ _ _/  ( ; -.    |    _ _\.-~        .-~~~^-.
  (( / / )) ,-{        _      `-.|.-~-.           .~         `.
 (( // / ))  '/\      /                 ~-. _ .-~      .-~^-.  \
 (( /// ))      `.   {            }                   /      \  \
  (( / ))     .----~-.\        \-'                 .~         \  `. \^-.
             ///.----..>        \             _ -~             `.  ^-`  ^-_
               ///-._ _ _ _ _ _ _}^ - - - - ~                     ~-- ,.-~
                                                                  /.-~

Cowsay.elephant_in_snake — Function

function elephant_in_snake()

Do we need to explain this?

Example

julia> cowsay("Help!", cow=Cowsay.elephant_in_snake)
 _______
< Help! >
 -------
   \
    \              ....
           ........    .
          .            .
         .             .
.........              .......
..............................

Elephant inside ASCII snake

Cowsay.eyes — Function

function eyes()

Evil-looking eyes

Example

julia> cowsay("Mwah, ha, ha!", cow=Cowsay.eyes)
 _______________
< Mwah, ha, ha! >
 ---------------
    \
     \
                                   .::!!!!!!!:.
  .!!!!!:.                        .:!!!!!!!!!!!!
  ~~~~!!!!!!.                 .:!!!!!!!!!UWWW$$$
      :$$NWX!!:           .:!!!!!!XUWW$$$$$$$$$P
      $$$$$##WX!:      .<!!!!UW$$$$"  $$$$$$$$#
      $$$$$  $$$UX   :!!UW$$$$$$$$$   4$$$$$*
      ^$$$B  $$$$\     $$$$$$$$$$$$   d$$R"
        "*$bd$$$$      '*$$$$$$$$$$$o+#"
             """"          """""""

Cowsay.flaming_sheep — Function

function flaming_sheep()

The flaming sheep, contributed by Geordan Rosario (geordan@csua.berkeley.edu)

Example

julia> cowsay("Baaahhh!", cow=Cowsay.flaming_sheep)
 __________
< Baaahhh! >
 ----------
  \            .    .     .
   \      .  . .     `  ,
    \    .; .  : .' :  :  : .
     \   i..`: i` i.i.,i  i .
      \   `,--.|i |i|ii|ii|i:
           UooU\.'@@@@@@`.||'
           \__/(@@@@@@@@@@)'
                (@@@@@@@@)
                `YY~~~~YY'
                 ||    ||

Cowsay.fox — Function

function fox()

Example

julia> cowsay("What does the fox say?", cow=Cowsay.fox)
 ________________________
< What does the fox say? >
 ------------------------
         \     ,-.      .-,
          \    |-.\ __ /.-|
           \   \  `    `  /
                /_     _ \
              <  _`q  p _  >
              <.._=/  \=_. >
                 {`\()/`}`\
                 {      }  \
                 |{    }    \
                 \ '--'   .- \
                 |-      /    \
                 | | | | |     ;
                 | | |.;.,..__ |
               .-"";`         `|
              /    |           /
              `-../____,..---'`

Cowsay.kitty — Function

function kitty()

A kitten of sorts, I think

Example

julia> cowsay("Meow", cow=Cowsay.kitty)
 ______
< Meow >
 ------
     \
      \
       ("`-'  '-/") .___..--' ' "`-._
         ` *_ *  )    `-.   (      ) .`-.__. `)
         (_Y_.) ' ._   )   `._` ;  `` -. .-'
      _.. `--'_..-_/   /--' _ .' ,4
   ( i l ),-''  ( l i),'  ( ( ! .-'

Cowsay.koala — Function

function koala

From the canonical koala collection

Example

julia> cowsay("Welcome to Down Under", cow=Cowsay.koala)
 _______________________
< Welcome to Down Under >
 -----------------------
  \
   \
       ___
     {~._.~}
      ( Y )
     ()~*~()
     (_)-(_)

Cowsay.meow — Function

function meow()

A meowing tiger?

Example

julia> cowsay("We're kind of stupid that way", cow=Cowsay.meow)
 _______________________________
< We're kind of stupid that way >
 -------------------------------
  \
   \ ,   _ ___.--'''`--''//-,-_--_.
      \`"' ` || \\ \ \\/ / // / ,-\\`,_
     /'`  \ \ || Y  | \|/ / // / - |__ `-,
    /@"\  ` \ `\ |  | ||/ // | \/  \  `-._`-,_.,
   /  _.-. `.-\,___/\ _/|_/_\_\/|_/ |     `-._._)
   `-'``/  /  |  // \__/\__  /  \__/ \
        `-'  /-\/  | -|   \__ \   |-' |
          __/\ / _/ \/ __,-'   ) ,' _|'
         (((__/(((_.' ((___..-'((__,'

Cowsay.moose — Function

function moose()

Example

julia> cowsay("Time to vamoose", cow=Cowsay.moose)
 _________________
< Time to vamoose >
 -----------------
  \
   \   \_\_    _/_/
    \      \__/
           (oo)\_______
           (__)\       )\/\
               ||----w |
               ||     ||

Cowsay.sheep — Function

function sheep()

The non-flaming sheep.

Example

julia> cowsay("Baa", cow=Cowsay.sheep)
 _____
< Baa >
 -----
  \
   \
       __
      UooU\.'@@@@@@`.
      \__/(@@@@@@@@@@)
           (@@@@@@@@)
           `YY~~~~YY'
            ||    ||

Cowsay.stegosaurus — Function

function stegosaurus()

A stegosaur with a top hat?

Example

julia> cowsay("How do you do?", cow=Cowsay.stegosaurus)
 ________________
< How do you do? >
 ----------------
\                             .       .
 \                           / `.   .' "
  \                  .---.  <    > <    >  .---.
   \                 |    \  \ - ~ ~ - /  /    |
         _____          ..-~             ~-..-~
        |     |   \~~~\.'                    `./~~~/
       ---------   \__/                        \__/
      .'  O    \     /               /       \  "
     (_____,    `._.'               |         }  \/~~~/
      `----.          /       }     |        /    \__/
            `-.      |       /      |       /      `. ,~~|
                ~-.__|      /_ - ~ ^|      /- _      `..-'
                     |     /        |     /     ~-.     `-. _  _  _
                     |_____|        |_____|         ~ - . _ _ _ _ _>

Cowsay.turkey — Function

function turkey()

Turkey!

Example

julia> cowsay("Gobble, gobble", cow=Cowsay.turkey)
 ________________
< Gobble, gobble >
 ----------------
  \                                  ,+*^^*+___+++_
   \                           ,*^^^^              )
    \                       _+*                     ^**+_
     \                    +^       _ _++*+_+++_,         )
              _+^^*+_    (     ,+*^ ^          \+_        )
             {       )  (    ,(    ,_+--+--,      ^)      ^\
            { (@)    } f   ,(  ,+-^ __*_*_  ^^\_   ^\       )
           {:;-/    (_+*-+^^^^^+*+*<_ _++_)_    )    )      /
          ( /  (    (        ,___    ^*+_+* )   <    <      \
           U _/     )    *--<  ) ^\-----++__)   )    )       )
            (      )  _(^)^^))  )  )\^^^^^))^*+/    /       /
          (      /  (_))_^)) )  )  ))^^^^^))^^^)__/     +^^
         (     ,/    (^))^))  )  ) ))^^^^^^^))^^)       _)
          *+__+*       (_))^)  ) ) ))^^^^^^))^^^^^)____*^
          \             \_)^)_)) ))^^^^^^^^^^))^^^^)
           (_             ^\__^^^^^^^^^^^^))^^^^^^^)
             ^\___            ^\__^^^^^^))^^^^^^^^)\\
                  ^^^^^\uuu/^^\uuu/^^^^\^\^\^\^\^\^\^\
                     ___) >____) >___   ^\_\_\_\_\_\_\)
                    ^^^//\\_^^//\\_^       ^(\_\_\_\)
                      ^^^ ^^ ^^^ ^

Cowsay.turtle — Function

function turtle()

A mysterious turtle...

Example

julia> cowsay("Where is that pesky rabbit?", cow=Cowsay.turtle)
 _____________________________
< Where is that pesky rabbit? >
 -----------------------------
    \                                  ___-------___
     \                             _-~~             ~~-_
      \                         _-~                    /~-_
             /^\__/^\         /~  \                   /    \
           /|  O|| O|        /      \_______________/        \
          | |___||__|      /       /                \          \
          |          \    /      /                    \          \
          |   (_______) /______/                        \_________ \
          |         / /         \                      /            \
           \         \^\\         \                  /               \     /
             \         ||           \______________/      _-_       //\__//
               \       ||------_-~~-_ ------------- \ --/~   ~\    || __/
                 ~-----||====/~     |==================|       |/~~~~~
                  (_(__/  ./     /                    \_\      \.
                         (_(___/                         \_____)_)

Cowsay.vader_koala — Function

function vader_koala

Another canonical koala?

Example

julia> cowsay("Luke, you are my joey!", cow=Cowsay.vader_koala)
 ________________________
< Luke, you are my joey! >
 ------------------------
   \
    \        .
     .---.  //
    Y|o o|Y//
   /_(i=i)K/
   ~()~*~()~
    (_)-(_)

     Darth
     Vader
     koala